*Info *Create ListDaftarPelamarComp *Create UpdatePertanyaanPelamar *rapihin ProfileComp

+++ BlazorWasmLoker/Resoruces/Users/PengalamanResourceUpdate.cs
@@ -11,10 +11,9 @@ namespace BlazorWasmLoker.Resoruces.Users
         public string TempatKerja { get; set; }
         public string Posisi { get; set; }
         public string Keterangan { get; set; }
-        public double Nominal { get; set; }
+        public decimal Nominal { get; set; }
         public DateTime TglAwal { get; set; }
         public DateTime TglAkhir { get; set; }
         public string masaKerja { get; set; }
-
     }
 }
